requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.13
django-cors-headers==3.2.1
django-extensions==2.1.3
django-filter==2.2.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-watchman==1.1.0
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
djangorestframework-recursive==0.1.2
gunicorn==19.9.0
idna==2.9
importlib-metadata==1.6.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.3
Pillow==7.1.1
psycopg2==2.8.5
pyparsing==2.4.6
pyrsistent==0.16.0
pytz==2018.9
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
static3==0.7.0
swagger-spec-validator==2.5.0
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.8
user-agents==2.1
whitenoise==5.0.1
zipp==3.1.0
django-admin-list-filter-dropdown==1.0.3
django-ckeditor==5.9.0
drf-flex-fields